pyproject.toml
[tool.poetry]
name = "norns"
version = "0.2.0"
description = ""
authors = ["Josh Odell <j01101111sh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
environs = {extras = ["django"], version = "^10.3.0"}
pytz = "^2024.1"
crispy-bootstrap5 = "^2024.2"
django-crispy-forms = "^2.1"
django-bootstrap-v5 = "^1.0.11"
whitenoise = "^6.6.0"
gunicorn = "^21.2.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django-allauth = "^0.61.1"
django-anymail = {extras = ["mailgun"], version = "^10.2"}
django-markdownx = "^4.0.7"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.7"
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"
pylint = "^3.1.0"
pyupgrade = "^3.15.1"
django-upgrade = "^1.16.0"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
factory-boy = "^3.3.0"
django-coverage-plugin = "^3.1.0"
pytest-django = "^4.8.0"
pytest-sugar = "^1.0.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
furo = "^2024.1.29"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311', 'py312']

[tool.isort]
profile = 'black'

[tool.pylint.main]
disable = [
    "R0801", # similar lines in 2 files
    "R0901", # too many ancestors
    "R0903", # too few public methods
    "C0415", # import outside toplevel
    "E1101", # no-member
]

[tool.pylint.format]
max-line-length = "88"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "norns.settings"
addopts = "--reuse-db"
python_files = "test_*.py"
python_functions = "test_*"

log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s - %(levelname)s - %(name)s:%(lineno)d\n\t%(message)s"

log_file_level = "DEBUG"
log_file_format = "%(asctime)s - %(levelname)s - %(name)s:%(lineno)d\n\t%(message)s"

[tool.coverage.report]
omit = ["tests/*.py"]
show_missing = true
